Modificiations to support solarwinds version of syslog
Solarwinds inserts an field to store the original IP address of the
syslog message. This is not valid according to any syslog specification.
Modify the schema to require this field, so this schema will only work
with messages forwarded by solarwinds.

Also moves the RFC5424 parsing to a separate choice so that we support
this and and other syslog variants.