Allow schema to validate using Xerces-C
Xerces-C has a bug where it cannot correctly validate choices with an
empty sequnce as one of the branches. This replaces the empty sequence
with an element parsing a string of length one that ensures the value is
a hyphen, which xerces-C does support. The downside is now the infoset
always constains a NoStructureData element, though there is no
workaround for this since Xerces-C needs an actual element in the
choice.

This method has been tested to be portable with IBM also be portable for
IBM DFDL. Additionally, xsi:type attributes were added to TDML infosets
to account for differences with timestamps representations in the
infoset with IBM DFDL. And fixed the portable priority value type to
correctly include the namespace prefix.